Relax edid-decode character checking
extract_string in edid-decode currently only returns characters that
pass isalnum(). This causes an error when trying to parse the name of my
A/V receiver, which is "HT-R791". The VESA E-DID standard states that
normal ASCII is allowed for this field, not just alphanumeric
characters. Patch below changes to using isgraph() which will allow
anything printable.

Signed-off-by: Jonathan McDowell <noodles@earth.li>

-----
-----

J.