edid-decode: add --extract and --check options
Use --extract to extract the contents of the first EDID block
as hex values. The was the default in old versions, use this
option to see this data again. I don't see the use of this
hex dump, so I turned it off by default.

If you want to check the EDID if it conforms to the standards,
then use the --check option.

So now the default behavior of edid-decode when run without these
options is to just parse the EDID and show it in human-readable text,
which is what you want.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>