edid-decode: make it easier to find the out-of-range monitor values
When the EDID is checked, it verifies if all the modes listed fit
within the ranges reported by the monitor. It warns if that is not
the case, but fails to show the actual monitor ranges.

By adding that information it is much easier to find which monitor
range needs to be extended.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>