edid-decode: parse additional flags in the DD+ Short Audio Descriptor
In the Dolby Digital Plus Short Audio Descriptor, byte 3 is used to
signal Joint Object Coding support, which can be useful to determine if
the sink supports Dolby Atmos.

Parse these flags as described in the Dolby Audio and Dolby Atmos over
HDMI specification v1.0. Other bits in byte 3 are marked as reserved and
should be zero.

Signed-off-by: Arnaud Vrac <avrac@freebox.fr>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>